requirements.txt
streamlit
python-dotenv
google-generativeai
langchain-community